refactor(ways): extract session::engagement module (ADR-123 cluster)
session.rs crossed the 800-line priority threshold during the PR #49
code-review response, landing at 809 after the backward-tick guard,
the classify_outcome split, and the FirstFire → ReFire → Suppressed
contract tests were added. The file's hottest sub-module boundary is
now along the ADR-123 firing-dynamics cluster, which is a coherent
ADR-123 unit that carries its own tests — a clean seam.

This commit moves the cluster into a new `session::engagement`
submodule. session.rs re-exports the public surface so every existing
caller (`session::REFIRE_FLOOR`, `session::way_fire_outcome`,
`session::FireOutcome`, etc.) keeps resolving with zero call-site
churn.

Contents moved:
- `REFIRE_FLOOR` const
- `FireOutcome` enum + impls (`is_allowed`, `is_redisclosure`)
- `engagement_path` / `load_engagement` / `save_engagement` (private)
- `way_fire_outcome` (pub)
- `classify_outcome` (private)
- `record_way_fire` (pub)
- `load_engagement_for_tick` (private)
- `way_refire_threshold_k` (pub)
- `fire_outcome_contract` + `load_engagement_for_tick_resets_on_backward_jump` tests

Line counts after the split:
  session.rs        609  (was 809 — back under 800 priority tier)
  session/engagement.rs  235

Cross-module wiring: engagement.rs reaches back into session.rs's
private helpers (`session_dir`, `ensure_parent`) via `super::`.
Rust's privacy model lets descendant modules see parent-private
items, so no visibility promotion was needed on the session.rs side.
`get_token_position` and `resolve_way_file` are already `pub`.

Closes #52.
